restruct filter expr
