+++ raphtory/src/core/utils/errors.rs
@@ -1,6 +1,6 @@
 use crate::{
     core::{storage::lazy_vec::IllegalSet, utils::time::error::ParseTimeError, Prop},
-    db::graph::views::filter::FilterOperator,
+    db::graph::views::filter::model::filter_operator::FilterOperator,
 };
 #[cfg(feature = "io")]
 use parquet::errors::ParquetError;
