+++ raphtory/src/db/api/storage/graph/storage_ops/time_semantics.rs
@@ -1250,7 +1250,7 @@ mod test_graph_storage {
     mod search_nodes {
         use super::*;
         use crate::{
-            db::{api::view::SearchableGraphOps, graph::views::filter::PropertyFilterOps},
+            db::{api::view::SearchableGraphOps, graph::views::filter::model::PropertyFilterOps},
             prelude::{Graph, NodeViewOps, PropertyFilter},
         };
 
@@ -1276,7 +1276,7 @@ mod test_graph_storage {
     mod search_edges {
         use super::*;
         use crate::{
-            db::{api::view::SearchableGraphOps, graph::views::filter::PropertyFilterOps},
+            db::{api::view::SearchableGraphOps, graph::views::filter::model::PropertyFilterOps},
             prelude::{EdgeViewOps, Graph, NodeViewOps, PropertyFilter},
         };
         #[test]
