+++ raphtory/src/db/graph/views/filter/edge_field_filtered_graph.rs
@@ -13,7 +13,7 @@ use crate::{
                 Base,
             },
         },
-        graph::views::filter::{internal::InternalEdgeFilterOps, EdgeFieldFilter, Filter},
+        graph::views::filter::{internal::InternalEdgeFilterOps, model::Filter, EdgeFieldFilter},
     },
     prelude::GraphViewOps,
 };
