+++ raphtory/src/db/graph/views/filter/edge_or_filtered_graph.rs
@@ -14,7 +14,7 @@ use crate::{
                 Base,
             },
         },
-        graph::views::filter::{internal::InternalEdgeFilterOps, OrFilter},
+        graph::views::filter::{internal::InternalEdgeFilterOps, model::OrFilter},
     },
     prelude::GraphViewOps,
 };
