+++ raphtory/src/db/graph/views/filter/model/filter_operator.rs
@@ -1,0 +1,172 @@
+use crate::{
+    core::Prop,
+    db::graph::views::filter::model::{property_filter::PropertyFilterValue, FilterValue},
+};
+use std::{collections::HashSet, fmt, fmt::Display, ops::Deref};
+use strsim::levenshtein;
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum FilterOperator {
+    Eq,
+    Ne,
+    Lt,
+    Le,
+    Gt,
+    Ge,
+    In,
+    NotIn,
+    IsSome,
+    IsNone,
+    Contains,
+    NotContains,
+    FuzzySearch {
+        levenshtein_distance: usize,
+        prefix_match: bool,
+    },
+}
+
+impl Display for FilterOperator {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        let operator = match self {
+            FilterOperator::Eq => "==",
+            FilterOperator::Ne => "!=",
+            FilterOperator::Lt => "<",
+            FilterOperator::Le => "<=",
+            FilterOperator::Gt => ">",
+            FilterOperator::Ge => ">=",
+            FilterOperator::In => "IN",
+            FilterOperator::NotIn => "NOT_IN",
+            FilterOperator::IsSome => "IS_SOME",
+            FilterOperator::IsNone => "IS_NONE",
+            FilterOperator::Contains => "CONTAINS",
+            FilterOperator::NotContains => "NOT_CONTAINS",
+            FilterOperator::FuzzySearch {
+                levenshtein_distance,
+                prefix_match,
+            } => {
+                return write!(f, "FUZZY_SEARCH({},{})", levenshtein_distance, prefix_match);
+            }
+        };
+        write!(f, "{}", operator)
+    }
+}
+
+impl FilterOperator {
+    pub fn is_strictly_numeric_operation(&self) -> bool {
+        matches!(
+            self,
+            FilterOperator::Lt | FilterOperator::Le | FilterOperator::Gt | FilterOperator::Ge
+        )
+    }
+
+    fn operation<T>(&self) -> impl Fn(&T, &T) -> bool
+    where
+        T: ?Sized + PartialEq + PartialOrd,
+    {
+        match self {
+            FilterOperator::Eq => T::eq,
+            FilterOperator::Ne => T::ne,
+            FilterOperator::Lt => T::lt,
+            FilterOperator::Le => T::le,
+            FilterOperator::Gt => T::gt,
+            FilterOperator::Ge => T::ge,
+            _ => panic!("Operation not supported for this operator"),
+        }
+    }
+
+    pub fn fuzzy_search(
+        &self,
+        levenshtein_distance: usize,
+        prefix_match: bool,
+    ) -> impl Fn(&str, &str) -> bool {
+        move |left: &str, right: &str| {
+            let left = left.to_lowercase();
+            let right = right.to_lowercase();
+            let levenshtein_match = levenshtein(&left, &right) <= levenshtein_distance;
+            let prefix_match = prefix_match && right.starts_with(&left);
+            levenshtein_match || prefix_match
+        }
+    }
+
+    fn collection_operation<T>(&self) -> impl Fn(&HashSet<T>, &T) -> bool
+    where
+        T: Eq + std::hash::Hash,
+    {
+        match self {
+            FilterOperator::In => |set: &HashSet<T>, value: &T| set.contains(value),
+            FilterOperator::NotIn => |set: &HashSet<T>, value: &T| !set.contains(value),
+            _ => panic!("Collection operation not supported for this operator"),
+        }
+    }
+
+    pub fn apply_to_property(&self, left: &PropertyFilterValue, right: Option<&Prop>) -> bool {
+        match left {
+            PropertyFilterValue::None => match self {
+                FilterOperator::IsSome => right.is_some(),
+                FilterOperator::IsNone => right.is_none(),
+                _ => unreachable!(),
+            },
+            PropertyFilterValue::Single(l) => match self {
+                FilterOperator::Eq
+                | FilterOperator::Ne
+                | FilterOperator::Lt
+                | FilterOperator::Le
+                | FilterOperator::Gt
+                | FilterOperator::Ge => right.map_or(false, |r| self.operation()(r, l)),
+                FilterOperator::Contains => right.map_or(false, |r| match (l, r) {
+                    (Prop::Str(l), Prop::Str(r)) => r.deref().contains(l.deref()),
+                    _ => unreachable!(),
+                }),
+                FilterOperator::NotContains => right.map_or(false, |r| match (l, r) {
+                    (Prop::Str(l), Prop::Str(r)) => !r.deref().contains(l.deref()),
+                    _ => unreachable!(),
+                }),
+                FilterOperator::FuzzySearch {
+                    levenshtein_distance,
+                    prefix_match,
+                } => right.map_or(false, |r| match (l, r) {
+                    (Prop::Str(l), Prop::Str(r)) => {
+                        let fuzzy_fn = self.fuzzy_search(*levenshtein_distance, *prefix_match);
+                        fuzzy_fn(l, r)
+                    }
+                    _ => unreachable!(),
+                }),
+                _ => unreachable!(),
+            },
+            PropertyFilterValue::Set(l) => match self {
+                FilterOperator::In | FilterOperator::NotIn => {
+                    right.map_or(false, |r| self.collection_operation()(l, r))
+                }
+                _ => unreachable!(),
+            },
+        }
+    }
+
+    pub fn apply(&self, left: &FilterValue, right: Option<&str>) -> bool {
+        match left {
+            FilterValue::Single(l) => match self {
+                FilterOperator::Eq | FilterOperator::Ne => match right {
+                    Some(r) => self.operation()(r, l),
+                    None => matches!(self, FilterOperator::Ne),
+                },
+                FilterOperator::Contains => right.map_or(false, |r| r.contains(l)),
+                FilterOperator::NotContains => right.map_or(false, |r| !r.contains(l)),
+                FilterOperator::FuzzySearch {
+                    levenshtein_distance,
+                    prefix_match,
+                } => right.map_or(false, |r| {
+                    let fuzzy_fn = self.fuzzy_search(*levenshtein_distance, *prefix_match);
+                    fuzzy_fn(l, r)
+                }),
+                _ => unreachable!(),
+            },
+            FilterValue::Set(l) => match self {
+                FilterOperator::In | FilterOperator::NotIn => match right {
+                    Some(r) => self.collection_operation()(l, &r.to_string()),
+                    None => matches!(self, FilterOperator::NotIn),
+                },
+                _ => unreachable!(),
+            },
+        }
+    }
+}
