+++ raphtory/src/db/graph/views/filter/node_name_filtered_graph.rs
@@ -13,7 +13,7 @@ use crate::{
                 Base,
             },
         },
-        graph::views::filter::{internal::InternalNodeFilterOps, Filter, NodeNameFilter},
+        graph::views::filter::{internal::InternalNodeFilterOps, model::Filter, NodeNameFilter},
     },
     prelude::GraphViewOps,
 };
